requirement.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
setuptools==80.9.0
SQLAlchemy==2.0.43
svgwrite==1.4.3
Tree==0.2.4
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1
